chore: moves test-utils.ts in __helpers__ dir
